layers: Special-case UO generation of QueuePresentKHR
VkPresentInfo::pResults is an /output/ array. The generated code makes a
copy on the way in (pointlessly, but other clients of safe_* might find
this reasonable), but neglects to copy it back on the way out.

Lift the current code generated for QueuePresentKHR back into the
special cases, and add the correct post-call copy.

Fixes #1670.

Signed-off-by: Chris Forbes <chrisforbes@google.com>